# src/main/java/com/technical/task/task_project/service/dto/AuthenticationRequestDTO.java
package com.technical.task.task_project.service.dto;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthenticationRequestDTO {
    @NotNull
    private String email;
    @NotNull
    private String password;
}